- added new class system in Class.js -- This system will allow for inheritance while not clogging the global namespace. I think our whole js should be under the namespace of Rootless, so that we will never run into problems with other libraries we might integrate. For the Map object on the page it will be under the namespace of Rootless.Map.
I also added a singleton pattern that will make sure never to allow more than one singleton of the same name in the DOM.

Added a Namespace function that will make sure to add each object into the "Rootless" namespace. I will make a map class that will serve as the first example of this. (I will try and submit that on Sunday)